Agents/requirements.txt
crewai
crewai-tools
python-dotenv>=1.0.0
flask>=2.0.0
flask-cors>=4.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
beautifulsoup4>=4.12.2
requests>=2.31.0
nest_asyncio>=1.5.8
alpha_vantage>=2.3.1
typing-extensions>=4.6.0
waitress>=2.1.2
PyPDF2>=3.0.0
werkzeug>=2.0.0


alpha_vantage
Flask
Flask-Cors
langchain
crewai
crewai-tools
